[tests] output unit test updates to go with broken out code
