+++ stream_alert/alert_processor/outputs/jira.py
@@ -40,7 +40,8 @@ class JiraOutput(OutputDispatcher):
         self._base_url = None
         self._auth_cookie = None
 
-    def get_user_defined_properties(self):
+    @classmethod
+    def get_user_defined_properties(cls):
         """Get properties that must be asssigned by the user when configuring a new Jira
         output.  This should be sensitive or unique information for this use-case that needs
         to come from the user.
