+++ stream_alert/alert_processor/outputs/output_base.py
@@ -45,7 +45,7 @@ class StreamAlertOutput(object):
         return output
 
     @classmethod
-    def get_output_dispatcher(cls, service, region, function_name, config):
+    def create_dispatcher(cls, service, region, function_name, config):
         """Returns the subclass that should handle this particular service
 
         Args:
@@ -57,8 +57,24 @@ class StreamAlertOutput(object):
         Returns:
             OutputDispatcher: Subclass of OutputDispatcher to use for sending alerts
         """
+        dispatcher = cls.get_dispatcher(service)
+        if not dispatcher:
+            return False
+
+        return dispatcher(region, function_name, config)
+
+    @classmethod
+    def get_dispatcher(cls, service):
+        """Returns the subclass that should handle this particular service
+
+        Args:
+            service (str): The service identifier for this output
+
+        Returns:
+            OutputDispatcher: Subclass of OutputDispatcher to use for sending alerts
+        """
         try:
-            return cls._outputs[service](region, function_name, config)
+            return cls._outputs[service]
         except KeyError:
             LOGGER.error('Designated output service [%s] does not exist', service)
 
@@ -74,7 +90,7 @@ class StreamAlertOutput(object):
 
 
 class OutputDispatcher(object):
-    """StreamOutputBase is the base class to handle routing alerts to outputs
+    """OutputDispatcher is the base class to handle routing alerts to outputs
 
     Public methods:
         get_secrets_bucket_name: returns the name of the s3 bucket for secrets that
@@ -305,7 +321,8 @@ class OutputDispatcher(object):
 
         return cred_name
 
-    def format_output_config(self, service_config, values):
+    @classmethod
+    def format_output_config(cls, service_config, values):
         """Add this descriptor to the list of descriptor this service
            If the service doesn't exist, a new entry is added to an empty list
 
@@ -315,10 +332,11 @@ class OutputDispatcher(object):
         Returns:
             [list<string>] List of descriptors for this service
         """
-        return service_config.get(self.__service__, []) + [values['descriptor'].value]
+        return service_config.get(cls.__service__, []) + [values['descriptor'].value]
 
+    @classmethod
     @abstractmethod
-    def get_user_defined_properties(self):
+    def get_user_defined_properties(cls):
         """Base method for retrieving properties that must be asssigned by the user when
         configuring a new output for this service. This should include any information that
         is sensitive or use-case specific. For intance, if the url needed for this integration
