+++ stream_alert/alert_processor/outputs/phantom.py
@@ -31,7 +31,8 @@ class PhantomOutput(OutputDispatcher):
     CONTAINER_ENDPOINT = 'rest/container'
     ARTIFACT_ENDPOINT = 'rest/artifact'
 
-    def get_user_defined_properties(self):
+    @classmethod
+    def get_user_defined_properties(cls):
         """Get properties that must be asssigned by the user when configuring a new Phantom
         output.  This should be sensitive or unique information for this use-case that needs
         to come from the user.
